Advanced-ATS-Resume-Checker/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
google-generativeai==0.3.2
gunicorn==21.2.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
pytesseract==0.3.10
pdfminer.six==20231228
python-docx==1.1.0
PyMuPDF==1.23.6
